arigato zedun tatu girls desu

+++ makoa/messages.hpp
@@ -27,7 +27,7 @@ static const uint32_t message_size = 48, message_bsize = message_size - 17;
 struct ttime_t
 {
     //value equals (unix_time * 10^9 + nanoseconds)
-    static const uint32_t frac = 1000000000;
+    static const uint64_t frac = 1000000000;
     uint64_t value;
 };
 #define TTIME_T_DEFINED
@@ -70,15 +70,15 @@ static_assert(sizeof(message_hello) == message_size, "protocol agreement");
 
 struct price_t
 {
-    static const int32_t exponent = -5;
-    static const uint32_t frac = 100000;
+    static const int64_t exponent = -5;
+    static const int64_t frac = 100000;
     int64_t value;
 };
 
 struct count_t
 {
-    static const int32_t exponent = -8;
-    static const uint32_t frac = 100000000;
+    static const int64_t exponent = -8;
+    static const int64_t frac = 100000000;
     int64_t value;
 };
